+++ bigmler/defaults.py
@@ -166,7 +166,8 @@ FLAGS = {
         {'flag': 'operating_point', 'type': 'string'},
         {'flag': 'step_out_of_bag', 'type': 'boolean'},
         {'flag': 'org_project', 'type': 'string'},
-        {'flag': 'split_field', 'type': 'string'}],
+        {'flag': 'split_field', 'type': 'string'},
+        {'flag': 'float_field', 'type': 'string'}],
     'BigMLer analyze': [
         {'flag': 'k-fold', 'type': 'int'},
         {'flag': 'cv', 'type': 'boolean'},